refactor: 패키지 이름 변경 및 클래스 이동
