+++ src/main/java/nbc/ticketing/ticket911/domain/seat/controller/SeatController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import nbc.ticketing.ticket911.application.seat.service.SeatService;
+import nbc.ticketing.ticket911.domain.seat.application.SeatService;
 import nbc.ticketing.ticket911.common.response.CommonResponse;
 import nbc.ticketing.ticket911.domain.seat.dto.request.CreateSeatRequestDto;
 import nbc.ticketing.ticket911.domain.seat.dto.request.UpdateSeatRequestDto;
