+++ src/main/java/nbc/ticketing/ticket911/domain/stage/controller/StageController.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import nbc.ticketing.ticket911.application.stage.service.StageService;
+import nbc.ticketing.ticket911.domain.stage.application.StageService;
 import nbc.ticketing.ticket911.common.response.CommonResponse;
 import nbc.ticketing.ticket911.domain.stage.dto.request.CreateStageRequestDto;
 import nbc.ticketing.ticket911.domain.stage.dto.request.UpdateStageRequestDto;
